Modifiche per classe:
-) Ordine : aggiunto l'attributo prezzo
-) OrdineDAO : aggiornati i metodi aggiungendo il prezzo dove serve
-) ControllerUtente : aggiunto il prezzo durante la creazione di un nuovo ordine

+++ src/Model/Servizio.java
@@ -57,7 +57,7 @@ public class Servizio {
         this.descrizione = descrizione;
     }
 
-    public double getPrezzo() {
+    public float getPrezzo() {
         return prezzo;
     }
